requirements.txt
plotly
pandas
psycopg2-binary
Flask
Werkzeug
